Organizing code and the drop of Calendar's methods
In order to have a more expressive API I decided to remove some of
Calendar's methods and preserve methods for date handling with calendar
support.
Now a Calendar instance is passed as a third parameter and some of the
most relevant bizdays methods handle only iso-formated dates.
With that change I expect I don't have check warnings and things like
that.