Calendar's adjustments arguments
Two new arguments have been added to `Calendar` function: `adjust.from` and `adjust.to`.
These arguments are used as filters for `bizdays` arguments `from` and `to`.
The default setting is `adjust.from=adjust.next` and `adjust.to=adjust.previous` and that reproduces the Excel's behavior.

The original issue was to implement these arguments directly in `bizdays` function, but implementing it in `Calendar` reinforces its portability.

#9