Calendar$add was changed to reflect the expected behavior
Calendar$add or offset has been changed to reflect the expected behavior that is: when a nonworking day is passed it offsets the given date according to the index.
So if 1 is added to this date, the next business day is returned, otherwise if -1 is added the previous business date is returned.

The test also have been rewritten.