Added the financial argument to create.calendar
The financial flag defines a financial calendar and this is the default option when creating a new calendar. To create non financial calendar the financial must be set to FALSE.

This feature allowed me to see that an adjust must be done in bizdays when from and to are, simultaneously, non bizdays.

issue #62